Cleaner detaching
insertAfter was creating cycles, by not setting next.prev to prev!

A better solution would be a setLink function, that if set to next, set the destination's next.prev.next to this, set the destination's next.prev to this, etc, to make sure no dangling links to old places were around. But... this should work...?